intermediate step with save bugs fixed
There were some errors in last commit to the bugfixing branch.
This are fixed here. now ready to work on getting the output from DISORT
suppressed.

+++ cia_arg_mod.f90
@@ -17,7 +17,7 @@ contains
     real,dimension(nciatemps,nwave)::ph2h2,ph2He,ph2h,ph2ch4
     real,dimension(4,nciatemps,nwave),intent(in):: cia
     real, dimension(nwave) :: ciah2h2, ciah2he,ciah2ch4    
-    double precision,dimension(nciatemps), intent(in) :: ciatemp 
+    real,dimension(nciatemps), intent(in) :: ciatemp 
     real, intent(in) :: grav
     integer, intent(in) :: ch4index
     real,dimension(nciatemps) :: tdiff
